pyproject.toml
[project]
name = "pdf-md-service"
version = "0.1.0"
description = "PDF to Markdown conversion service using Marker and Ray Serve"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.7",
    "httpx>=0.28.1",
    "uvicorn>=0.40.0",
    "ray[serve]>=2.40",
    "marker-pdf==1.8.2",
    "torch==2.7.1",
    "openai>=1.50",
    "python-multipart>=0.0.9",
]
